🐞 Prevent the runnable bits from modules to run on compiled lib
When running the compiled library as a script, every bit of code that
each module decides to run when is not being sourced would run here.

This is not desireable for the full compiled library.

Now modules must use meta::module_is_running, which will additionally
check if a build date is set (only true for the compiled library).

This also makes necessary to include the functions in _meta.sh into the
compiled library.